requirements.txt
torch==2.0.0
numpy==1.24.0
Pillow==9.4.0
scipy==1.10.0
h5py
joblib
gradio
tqdm
scikit-learn